AAI520_Market_Mover_Daily_Brief_Agent/requirements.txt
# Core Dependencies
yfinance==0.2.33
requests==2.31.0
python-dotenv==1.0.0
pandas==1.5.3  # Using an older version with pre-built wheels
numpy==1.24.3
python-dateutil==2.8.2
beautifulsoup4==4.12.2
lxml==4.9.3
newsapi-python==0.2.7
plotly==5.18.0
pandas-ta==0.3.14b0

# Web Framework
Flask==2.3.3
Flask-SocketIO==5.3.5
eventlet==0.33.3
python-socketio==5.10.0

# Data Processing
scikit-learn==1.3.0
nltk==3.8.1

# WebSockets
websocket-client==1.6.1
python-engineio==4.8.2

# Async
gevent==23.7.0
gevent-websocket==0.10.1

# Other
bidict==0.23.1
